(copy_internal): Don't explicitly unlink the destination when moving a symlink into the place of an existing non-directory. Reported by Joshua Hudson.

+++ src/copy.c
@@ -1230,8 +1230,10 @@ copy_internal (char const *src_name, char const *dst_name,
 	  else if (! S_ISDIR (dst_sb.st_mode)
 		   && (x->unlink_dest_before_opening
 		       || (x->preserve_links && 1 < dst_sb.st_nlink)
-		       || (x->dereference == DEREF_NEVER
-			   && S_ISLNK (src_sb.st_mode))))
+		       || (!x->move_mode
+			   && x->dereference == DEREF_NEVER
+			   && S_ISLNK (src_sb.st_mode))
+		       ))
 	    {
 	      if (unlink (dst_name) != 0 && errno != ENOENT)
 		{
